move files to accounting module directory
